Refactoring and life bar implementation
Some code was refactor, some debug logs were removed and some comments,
also. The hungry bar scripts were added and joined with a new one which
controls the decreasing of the hungry bar. Despite this, the script is
not completed yet since it must linked to the gamemanager and a game
over state.